Simplified APU's ExecuteFrameCounter
Added TimesTwo bool field to Divider which automatically multiplies the
divider's period by two.  This is used in the triangle and pulse
channel's dividers.  This fixes the issue where the pitch was too high
for these channels.

NewRP2A03 takes an apuFrequency which used to set the targetCycles in
NewAPU which determines how frequently the APU outputs a sample to the
DAC.

Removed NES.Save/LoadState and the accompanying events since they didn't
work anyways.