Added a simple debug web console
Run nintengo with -http=':6060' to run a web server on localhost:6060,
open it in a web browser to see the current state of variables/registers
in the CPU/DMA/APU/PPU/OAM etc.  Also displays a hexdump of the CPU and
PPU's memory space as well as the current state of the PPU pattern
tables as PNG images.  Refresh anytime to get current values.