Add encoding/json tags to Registers/Memory fields in m65go2.M6502, rp2ago3.RP2A03 and rp2ago3.APU.  Because m65go2.M6502 and rp2ago3.APU are embedded in rp2ago3.RP2A03, we need to give unique names to their Registers/Memory fields or else encoding/json will only include the last.  This was causing some save states to hang/crash the emulator on load because the CPU registers were never actually recorder in the save state file.  The save state version in meta.json has been bumped to 0.2.
Wrap NES.runProcessors in an anonymous function so we can print any
errors it returns and call os.Exit

Added SetSpeed(float) method to Audio interface.  This isn't totally
working yet, but will be needed so that Azul3DAudio can adjust it's
frequency when the emulator is slowed down using F9-F12.