nes: fix data race with nes.FrameEvent.Colors slice
Because video handling runs separately, the video handling (like azul3d_video.go)
will assume that the `colors` slice is still valid, but in reality the caller to
(*NES).frame assumes that once it returns they can continue modfiying it for the
next frame. This creates a data race where the colors might be invalid (detected
by Go race detector).