nes: fix data race
When `card` is passed to `Draw` ownership of it is transferred _to the device_,
after which ownership is only given back once `Render` returns. By allowing the
card to be unlocked inbetween the `Draw` and `Render` timeperiod, we are enabling
a data race on the texture.

This change is correct, however it makes the application unbearably slow. This
will be fixed in a follow-up commit by making the code more idiomatic.