nes: communicate texture updates via a channel
This change has `updateTex` communicate updates to the texture via a
channel to the render loop instead of using a lock. This improves the
performance of the prior locking solution drastically, because the
lock itself does not need to be held during `Render` which is tied to
the 60hz monitor refresh rate.