nes: do not needlessly set Clock.SetMaxFrameRate
At the very best, this is by default 75FPS and doesn't affect anything
because of VSync anyway. At the very worst, setting it to the same as
desired framerate (and not higher) [could be harmful](https://github.com/azul3d/engine/issues/154).

The default is probably always good enough for Nintengo.