nes: run azul3d_video event processing in a separate goroutine
By using `Poll` we only process events within the render loop, i.e., when
all rendering is actively occuring. When nintengo is paused, `<-texUpdate`
blocks and causes the render loop to halt. Because it doesn't reach the
next frame while blocking on `<-texUpdate`, `Poll` doesn't run again and
thus nintengo cannot be unpaused! Run event handling in a separate goroutine
to fix this.